Program structure mockup
This is the (incomplete) plan of the structure.

go to excalidraw.com and open the file.
Along with the flowchart, important discussions and important info to consider will be in here.

I hope to finish the planning for the program by today so we can all have a clear idea of how the program is going to look like asap cuz we only got 9 days. (the 30th doesn't count.)