Project File structure reworked
Everything's basically the same except, rather than grouping by task folders with different function files ,

functions are grouped by a file.c named after a task.

Header files have been created but they are blank. Need definition.

This lightens the amount of imports needed to go into main.c

project standards.md need to be updated to match.

This branch is to be made the default branch from now on so we have the main branch as a backup in case anything goes wrong in here.